refactor: move `MultiDimDataPage{Config,Url}` to `utils`
